tweaking .travis.yml; fixes for gcc 4.6

+++ cxx/tomographer2/histogram.h
@@ -76,7 +76,7 @@ struct UniformBinsHistogramParams
   //! Copy constructor, from any other UniformBinsHistogram::Params type.
   template<typename Params2,
            // enforce Params-like type by checking that properties 'min','max','num_bins' exist:
-           decltype((int)(Params2().min + Params2().max + Params2().num_bins)) dummyval = 0>
+           decltype((int)Params2().min + (int)Params2().max + Params2().num_bins) dummyval = 0>
   UniformBinsHistogramParams(const Params2& other)
     : min(other.min), max(other.max), num_bins(other.num_bins)
   {
@@ -210,7 +210,7 @@ struct UniformBinsHistogram
   //! Constructor: stores the parameters and initializes the histogram to zero counts everywhere
   template<typename Params2 = Params,
            // enforce Params-like type by checking that properties 'min','max','num_bins' exist:
-           decltype((int)(Params2().min + Params2().max + Params2().num_bins)) dummyval = 0>
+           decltype((int)Params2().min + (int)Params2().max + Params2().num_bins) dummyval = 0>
   UniformBinsHistogram(Params2 p = Params())
     : params(p), bins(Eigen::Array<CountType,Eigen::Dynamic,1>::Zero(p.num_bins)),
       off_chart(0)
